Move the services interfaces to the /Services folder
